Added code to see if King is being attacked

+++ forge/RandomSolver.h
@@ -13,6 +13,6 @@ namespace forge
 		virtual Move getMove(const Position & position) override;
 
 	protected:
-		std::default_random_engine m_randomEngine{ time(nullptr) };
+		std::default_random_engine m_randomEngine{ static_cast<unsigned int>(time(nullptr)) };
 	};
 } // namespace forge
